tripleLogger Reformatting, Added Comments
Added a forth logger that tracks temperature, humidity, and Mayfly stats (basically, anything that is not soil gas data).

Changed/added some comments for  relevance and general readability.

I'm now having to work with an older version of the Mayfly board, so another line was added at 120 to make it easy to switch between board versions.

This update has been tested, and appears to work as expected with all the listed sensors.